requirements.txt
sanic==23.3.0
loguru==0.6.0
matplotlib~=3.9.2
numpy==2.1.3
scipy==1.14.1
pandas==2.2.3
scikit-learn==1.5.2
statsmodels~=0.14.3
flake8==7.1.1
toml~=0.10.2
PyMySQL~=1.1.1
SQLAlchemy~=2.0.36
records==0.5.3
redis==4.4.0
pyodps==0.11.6.5
schedule==1.2.2
requests_toolbelt==1.0.0
requests==2.32.3
tenacity==9.0.0